requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.0.1
crashtest==0.4.1
cryptography==41.0.4
distlib==0.3.7
Django==4.2.6
django-filter==23.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
dulwich==0.21.6
filelock==3.12.2
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
jaraco.classes==3.3.0
jeepney==0.8.0
jsonschema==4.17.3
keyring==24.2.0
more-itertools==10.1.0
msgpack==1.0.7
packaging==23.2
pexpect==4.8.0
Pillow==10.0.1
pipenv==2023.10.3
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.8.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3.post1
rapidfuzz==2.15.1
requests==2.31.0
requests-toolbelt==1.0.0
SecretStorage==3.3.3
shellingham==1.5.3
sqlparse==0.4.4
tomlkit==0.12.1
trove-classifiers==2023.9.19
urllib3==2.0.6
virtualenv==20.24.5
zipp==3.17.0
